Plugin: Google Two Factor Authentication
This plugin allows Agents to use the Google Authenticator application
on their phone for 2FA. If configured, the Agent must scan a QR code
from the app, and when they go to log into the helpdesk, they will be
prompted to enter a 6-digit code displayed in the app to finish logging
in.

Note: This branch is specific to be used for SupportSystem. The code
for modules would need to be pushed in order to use the plugin.